# Conflicts:
#	src/cargopacket.cpp
#	src/linkgraph/linkgraphjob.cpp
#	src/newgrf_station.cpp
#	src/order_cmd.cpp
#	src/order_type.h
#	src/saveload/station_sl.cpp
#	src/script/api/script_road.hpp
#	src/script/api/script_stationlist.cpp
#	src/script/api/script_waypointlist.cpp
#	src/sl/oldloader_sl.cpp
#	src/spritecache.cpp
#	src/station_cmd.cpp
#	src/station_map.h
#	src/station_type.h
#	src/train_cmd.cpp
#	src/vehicle_base.h

+++ src/cachecheck.cpp
@@ -215,7 +215,7 @@ void CheckCaches(bool force_check, std::function<void(std::string_view)> log, Ch
 				cclog("station industries_near mismatch: st {}, (old size: {}, new size: {})", st->index, (uint)old_station_industries_nears[i].size(), (uint)st->industries_near.size());
 			}
 			if (!(old_station_catchment_tiles[i] == st->catchment_tiles)) {
-				cclog("station catchment_tiles mismatch: st {}", (int)st->index);
+				cclog("station catchment_tiles mismatch: st {}", st->index);
 			}
 			if (!(old_station_tiles[i] == st->station_tiles)) {
 				cclog("station station_tiles mismatch: st {}, (old: {}, new: {})", st->index, old_station_tiles[i], st->station_tiles);
